Combine document-1 and document-2, but serve them under two URLs for static and dynamic perspective
