+++ src/api/open-resource-discovery/v1/data/configuration.ts
@@ -4,27 +4,15 @@ import { customAccessStrategyLocalTenantId, customAccessStrategyGlobalTenantId, 
 export const ordConfiguration: ORDConfiguration = {
   openResourceDiscoveryV1: {
     documents: [
-      // Document 1 only contains static metadata
+      // Serve static metadata with open access strategy, ignore tenant headers
       {
-        url: '/open-resource-discovery/v1/documents/1',
+        url: '/open-resource-discovery/v1/documents/system-version',
         accessStrategies: [openAccessStrategy],
         perspective: 'system-version',
       },
-      // The system-instance perspective is identical to the system-version perspective
-      // So we just link to the same document again
+      // Serve dynamic metadata, requires system / tenant headers and the correct access strategy
       {
-        url: '/open-resource-discovery/v1/documents/1',
-        accessStrategies: [customAccessStrategyGlobalTenantId, customAccessStrategyLocalTenantId],
-        perspective: 'system-instance',
-      },
-      // Document 2 contains dynamic metadata, but only if accessed with a system / tenant header and the correct access strategy
-      {
-        url: '/open-resource-discovery/v1/documents/2',
-        accessStrategies: [openAccessStrategy],
-        perspective: 'system-version',
-      },
-      {
-        url: '/open-resource-discovery/v1/documents/2',
+        url: '/open-resource-discovery/v1/documents/system-instance',
         accessStrategies: [customAccessStrategyGlobalTenantId, customAccessStrategyLocalTenantId],
         perspective: 'system-instance',
       },
