+++ src/api/shared/validateUserAuthorization.ts
@@ -44,12 +44,19 @@ export function getTenantIdsFromHeader(req: FastifyRequest): {
   localTenantId: string | undefined
   sapGlobalTenantId: string | undefined
 } {
-  const localTenantId = _.isArray(req.headers['sap-local-tenant-id'])
-    ? req.headers['sap-local-tenant-id'].join()
-    : req.headers['sap-local-tenant-id']
-  const sapGlobalTenantId = _.isArray(req.headers['sap-global-tenant-id'])
-    ? req.headers['sap-global-tenant-id'].join()
-    : req.headers['sap-global-tenant-id']
+  let localTenantId = _.isArray(req.headers['local-tenant-id'])
+    ? req.headers['local-tenant-id'].join()
+    : req.headers['local-tenant-id']
+  let sapGlobalTenantId = _.isArray(req.headers['global-tenant-id'])
+    ? req.headers['global-tenant-id'].join()
+    : req.headers['global-tenant-id']
+
+  if (req.query && (req.query as Record<string, string>)['local-tenant-id']) {
+    localTenantId = (req.query as Record<string, string>)['local-tenant-id']
+  }
+  if (req.query && (req.query as Record<string, string>)['global-tenant-id']) {
+    sapGlobalTenantId = (req.query as Record<string, string>)['global-tenant-id']
+  }
 
   return {
     localTenantId,
